Create User by Admin done

+++ backend/src/Controllers/TourController.php
@@ -63,42 +63,63 @@ class TourController {
                 }
                 break;
             case 'POST':
-                if(isset($this->data->tours)) {
-                    $this->tour->from_city = $this->data->tours->from;
-                    $this->tour->to_city = $this->data->tours->to;
+                if(Validator::isAdmin() || Validator::isSuper()) {
+                    if(isset($this->data->tours)) {
+                        $this->tour->from_city = $this->data->tours->from;
+                        $this->tour->to_city = $this->data->tours->to;
+                        $this->tour->departures = $this->data->tours->departures;
+                        $this->tour->time = $this->data->tours->time;
+                        $this->tour->duration = $this->data->tours->duration;
+                        $this->tour->price = $this->data->tours->price;
+                        $this->tour->seats = $this->data->tours->seats;
+                        $this->tour->create();
+                    }
+                } else {
+                    http_response_code(403);
+                    echo json_encode([
+                        'error'=> 'Niste autorizovani da dodajete ture!'
+                    ], JSON_UNESCAPED_UNICODE);
+                }
+                break;
+            case 'PUT':
+                if(Validator::isAdmin() || Validator::isSuper()) {
+                    $this->tour->id = $this->data->tours->id;
                     $this->tour->departures = $this->data->tours->departures;
                     $this->tour->time = $this->data->tours->time;
                     $this->tour->duration = $this->data->tours->duration;
                     $this->tour->price = $this->data->tours->price;
                     $this->tour->seats = $this->data->tours->seats;
-                    $this->tour->create();
-                }
-                break;
-            case 'PUT':
-                $this->tour->id = $this->data->tours->id;
-                $this->tour->departures = $this->data->tours->departures;
-                $this->tour->time = $this->data->tours->time;
-                $this->tour->duration = $this->data->tours->duration;
-                $this->tour->price = $this->data->tours->price;
-                $this->tour->seats = $this->data->tours->seats;
-                if(isset($this->data->tours->update)) {
-                    $this->tour->update();
+                    if(isset($this->data->tours->update)) {
+                        $this->tour->update();
+                    }
+                } else {
+                    http_response_code(403);
+                    echo json_encode([
+                        'error'=> 'Niste autorizovani da uređujete ture!'
+                    ], JSON_UNESCAPED_UNICODE);
                 }
                 break;
             case 'DELETE':
-                if(isset($this->data->tours->id)) {
-                    $this->tour->id = $this->data->tours->id;
-                    $this->tour->to_city = $this->data->tours->to_city ?? null;
-                }
-                
-                if(isset($this->data->tours->delete)) {
-                    $this->tour->delete();
-                } elseif(isset($this->data->tours->restore)) {
-                    $this->tour->restore();
-                } elseif(isset($this->data->tours->restoreAll)) {
-                    $this->tour->restoreAll();
-                } elseif(isset($this->data->tours->permanentDelete)) {
-                    $this->tour->permanentDelete();
+                if(Validator::isAdmin() || Validator::isSuper()) {
+                    if(isset($this->data->tours->id)) {
+                        $this->tour->id = $this->data->tours->id;
+                        $this->tour->to_city = $this->data->tours->to_city ?? null;
+                    }
+                    
+                    if(isset($this->data->tours->delete)) {
+                        $this->tour->delete();
+                    } elseif(isset($this->data->tours->restore)) {
+                        $this->tour->restore();
+                    } elseif(isset($this->data->tours->restoreAll)) {
+                        $this->tour->restoreAll();
+                    } elseif(isset($this->data->tours->permanentDelete)) {
+                        $this->tour->permanentDelete();
+                    }
+                } else {
+                    http_response_code(403);
+                    echo json_encode([
+                        'error'=> 'Niste autorizovani da uređujete ture!'
+                    ], JSON_UNESCAPED_UNICODE);
                 }
                 break;
         }
